Moved WeatherModel into Model
It is now possible to use the location received from the browser and get
the weather forecast for that location. Untill we have implemented a way
to select a date the information received is currently only a forecast
for the week. This will of course change once we are able to pick a
specific day.