Firebase is alive!!! :fire: :baseball:
It works, but damn is it dodgy. A lot of things were done in a haste
because we need to finish this thing and if it works - it works. Does it
break? Probably.. But we need to meet the requirements before the
deadline and so this is a necessary evil? Anyway.. Firebase!