Started with remove activity :crossed_flags:
Tried to make remove activity work - but don't know how to get
information about which day the activity is in, only the index. If I
either can get the activity object or the day index we can implement
this function. Till then I removed the remove button since we don't
support it right now. If anyone knows how angular can pass along the
object used in the list I'd love to know.